# src/IO.Swagger/Controllers/AuthenticationApi.cs
/*
 * Calculator API
 *
 * A REST API for performing arithmetic operations. 
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using IO.Swagger.Exceptions;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;
using IO.Swagger.Services;
using Serilog;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class AuthenticationApiController : ControllerBase
    {
        private readonly IAuthService _authService;
        private readonly ILogger _logger;

        /// <summary>
        /// Initializes a new instance of the AuthenticationApiController
        /// </summary>
        /// <param name="authService">Authentication service</param>
        /// <param name="logger">Logger instance</param>
        public AuthenticationApiController(IAuthService authService, ILogger logger)
        {
            _authService = authService;
            _logger = logger;
        }

        /// <summary>
        /// Authenticate user and obtain JWT token
        /// </summary>
        /// <param name="body"></param>
        /// <response code="200">Authentication successful</response>
        /// <response code="401">Invalid credentials</response>
        /// <response code="409">Conflict - Username already taken</response>
        [HttpPost]
        [Route("/v1/auth/login")]
        [ValidateModelState]
        [SwaggerOperation("Login")]
        [SwaggerResponse(statusCode: 200, type: typeof(LoginResponse), description: "Authentication successful")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Invalid credentials")]
        [SwaggerResponse(statusCode: 409, type: typeof(Error), description: "Conflict - Username already taken")]
        public async Task<IActionResult> Login([FromBody] LoginRequest body)
        { 
            _logger.Information("Login/Register request received for username: {Username}", body?.Username);
            
            try
            {
                var result = await _authService.LoginOrRegisterAsync(body);
                _logger.Information("Authentication successful for username: {Username}", body.Username);
                return Ok(result);
            }
            catch (InvalidPasswordException ex)
            {
                // Invalid password - return 401 Unauthorized
                _logger.Warning("Invalid password attempt for username: {Username}", body?.Username);
                return Unauthorized(new Error
                {
                    _Error = "Invalid Credentials",
                    Message = ex.Message,
                    Timestamp = DateTime.UtcNow
                });
            }
            catch (ArgumentException ex)
            {
                // Invalid input - return 400 Bad Request
                _logger.Warning("Invalid input for authentication: {Message}", ex.Message);
                return BadRequest(new Error
                {
                    _Error = "Invalid Input",
                    Message = ex.Message,
                    Timestamp = DateTime.UtcNow
                });
            }
            catch (Exception ex)
            {
                // Unexpected errors - return 500 Internal Server Error
                _logger.Error(ex, "Unexpected error during authentication for username: {Username}", body?.Username);
                return StatusCode(500, new Error
                {
                    _Error = "Internal Server Error",
                    Message = "An unexpected error occurred during authentication.",
                    Timestamp = DateTime.UtcNow
                });
            }
        }
    }
}