# src/IO.Swagger/Controllers/CalculationsApi.cs
/*
 * Calculator API
 *
 * A REST API for performing arithmetic operations. 
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Common;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Authentication.JwtBearer;
using Serilog;
using IO.Swagger.Models;
using IO.Swagger.Services;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CalculationsApiController : ControllerBase
    {
        private readonly ICalculatorService _calculatorService;
        private readonly ILogger _logger;

        /// <summary>
        /// Initializes a new instance of the CalculationsApiController
        /// </summary>
        /// <param name="calculatorService">Calculator service</param>
        /// <param name="logger">Logger instance</param>
        public CalculationsApiController(ICalculatorService calculatorService, ILogger logger)
        {
            _calculatorService = calculatorService;
            _logger = logger;
        }

        /// <summary>
        /// Perform an arithmetic operation (Authorized)
        /// </summary>
        /// <remarks>Performs mathematical operations. Requires a valid JWT.</remarks>
        /// <param name="body"></param>
        /// <param name="xOperation"></param>
        /// <response code="200">Success</response>
        /// <response code="400">Invalid input or division by zero</response>
        /// <response code="401">Missing or invalid JWT token</response>
        [HttpPost]
        [Route("/v1/calculate")]
        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
        [ValidateModelState]
        [SwaggerOperation("PerformCalculation")]
        [SwaggerResponse(statusCode: 200, type: typeof(CalculationResponse), description: "Success")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Invalid input or division by zero")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Missing or invalid JWT token")]
        public async Task<IActionResult> PerformCalculation(
            [FromBody] CalculationRequest body, 
            [FromHeader(Name = "X-Operation"), Required] OperationType operation)
        { 
            _logger.Information("Calculation request received with operation: {Operation}", operation);
            
            try
            {
                var result = await _calculatorService.CalculateAsync(body, operation);
                _logger.Information("Calculation completed successfully");
                return Ok(result);
            }
            catch (ArgumentException ex)
            {
                // Invalid operation or input validation errors
                _logger.Warning("Invalid input for calculation: {Message}", ex.Message);
                return BadRequest(new Error
                {
                    _Error = "Invalid Input",
                    Message = ex.Message,
                    Timestamp = DateTime.UtcNow
                });
            }
            catch (InvalidOperationException ex)
            {
                // Division by zero or overflow errors
                _logger.Warning("Calculation error: {Message}", ex.Message);
                return BadRequest(new Error
                {
                    _Error = "Calculation Error",
                    Message = ex.Message,
                    Timestamp = DateTime.UtcNow
                });
            }
            catch (Exception ex)
            {
                // Unexpected errors
                _logger.Error(ex, "Unexpected error during calculation");
                return StatusCode(500, new Error
                {
                    _Error = "Internal Server Error",
                    Message = "An unexpected error occurred during calculation.",
                    Timestamp = DateTime.UtcNow
                });
            }
        }
    }
}